move NodeInfo to models folder
