+++ src/app/common/directives/node-delete.directive.ts
@@ -27,7 +27,8 @@ import { Directive, HostListener, Input } from '@angular/core';
 import { MinimalNodeEntity } from 'alfresco-js-api';
 import { Store } from '@ngrx/store';
 import { AppStore } from '../../store/states/app.state';
-import { DeleteNodesAction, NodeInfo } from '../../store/actions';
+import { DeleteNodesAction } from '../../store/actions';
+import { NodeInfo } from '../../store/models';
 
 @Directive({
     selector: '[acaDeleteNode]'
