+++ src/app/store/models.ts
@@ -1,0 +1,1 @@
+export * from './models/node-info.model';
