+++ src/app/store/models/node-info.model.ts
@@ -1,0 +1,7 @@
+export interface NodeInfo {
+    parentId?: string;
+    id: string;
+    name: string;
+    isFile?: boolean;
+    isFolder?: boolean;
+}
